+++ celements-model/src/main/java/com/celements/store/id/CelementsIdComputer.java
@@ -1,4 +1,4 @@
-package com.celements.model.doc;
+package com.celements.store.id;
 
 import javax.annotation.Nullable;
 import javax.validation.constraints.NotNull;
